refactor(src/common/containers): rename UsersItem to UserItem
